Fix #35: Use `AC_HEADER_TIME` and fix warning

+++ plugins/txtout/txtout.c
@@ -32,6 +32,8 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
+#include "config.h"
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -162,7 +164,7 @@ txtout_output(const char *descr, iaddr from, iaddr to, uint8_t proto, unsigned f
 	/*
 	 * IP Stuff
 	 */
-	fprintf(out, "%10ld.%06ld", ts.tv_sec, ts.tv_usec);
+	fprintf(out, "%10ld.%06ld", ts.tv_sec, (long)ts.tv_usec);
 	fprintf(out, " %s %u", ia_str(from), sport);
 	fprintf(out, " %s %u", ia_str(to), dport);
 	fprintf(out, " %hhu", proto);
